completed 23 & 24

+++ ca.mcgill.ecse223.kingdomino/src/main/java/ca/mcgill/ecse223/kingdomino/controller/KingdominoController.java
@@ -1252,12 +1252,7 @@ public class KingdominoController {
 				}
 			}
 		}
-		System.out.print((maxY - minY) < 5);
-		System.out.print(maxY);
-		System.out.print(minY);
-		System.out.print((maxX - minX) < 5);
-		System.out.print(maxX);
-		System.out.print(minX);
+		
 		return ((maxX - minX) < 5 && (maxY - minY) < 5);
 	}
 
@@ -1350,7 +1345,7 @@ public class KingdominoController {
 		Game game = kingdomino.getCurrentGame();
 		Player player = game.getNextPlayer();
 		Kingdom kingdom = player.getKingdom();
-
+		
 		List<KingdomTerritory> territories = new ArrayList<KingdomTerritory>(kingdom.getTerritories());
 
 		ArrayList<HashMap<String, Object>> tiles = new ArrayList<>();
@@ -1465,7 +1460,7 @@ public class KingdominoController {
 				// top, a tile on the left)
 				// or a new tile was added between three tiles (e.g. a tile on top, a tile on
 				// the left, a tile on the right)
-
+				
 				boolean myLeft = l && tiles.get(locationl).get("property") != null
 						&& tile.get("terrain").equals(tiles.get(locationl).get("terrain"));
 				boolean myUp = u && tiles.get(locationu).get("property") != null
@@ -1474,70 +1469,26 @@ public class KingdominoController {
 						&& tile.get("terrain").equals(tiles.get(locationr).get("terrain"));
 				boolean myDown = d && tiles.get(locationd).get("property") != null
 						&& tile.get("terrain").equals(tiles.get(locationd).get("terrain"));
-				if (myLeft && myRight) {
-					tile.put("property", tiles.get(locationl).get("property"));
-					Property prop = (Property) tiles.get(locationl).get("property");
-					prop.addIncludedDomino((Domino) tile.get("domino"));
-					if (!tiles.get(locationr).get("property").equals(tiles.get(locationl).get("property"))) {
-						Property tem = (Property) tiles.get(locationr).get("property");
-						tem.delete();
-						tiles.get(locationr).put("property", tiles.get(locationl).get("property"));
-						prop.addIncludedDomino((Domino) tiles.get(locationr).get("domino"));
-
-					}
-				}
-				if (myLeft && myUp) {
-					tile.put("property", tiles.get(locationl).get("property"));
-					Property prop = (Property) tiles.get(locationl).get("property");
-					prop.addIncludedDomino((Domino) tile.get("domino"));
-					if (!tiles.get(locationu).get("property").equals(tiles.get(locationl).get("property"))) {
-						Property tem = (Property) tiles.get(locationu).get("property");
-						tem.delete();
-						tiles.get(locationu).put("property", tiles.get(locationl).get("property"));
-						prop.addIncludedDomino((Domino) tiles.get(locationu).get("domino"));
-
-					}
-				}
-				if (myLeft && myDown) {
-					tile.put("property", tiles.get(locationl).get("property"));
-					Property prop = (Property) tiles.get(locationl).get("property");
-					prop.addIncludedDomino((Domino) tile.get("domino"));
-					if (!tiles.get(locationd).get("property").equals(tiles.get(locationl).get("property"))) {
-						Property tem = (Property) tiles.get(locationd).get("property");
-						tem.delete();
-						tiles.get(locationd).put("property", tiles.get(locationl).get("property"));
-						prop.addIncludedDomino((Domino) tiles.get(locationd).get("domino"));
-
-					}
-				}
-				if (myUp && myRight) {
-					tile.put("property", tiles.get(locationu).get("property"));
-					Property prop = (Property) tiles.get(locationu).get("property");
-					prop.addIncludedDomino((Domino) tile.get("domino"));
-					if (!tiles.get(locationr).get("property").equals(tiles.get(locationu).get("property"))) {
-						Property tem = (Property) tiles.get(locationr).get("property");
-						tem.delete();
-						tiles.get(locationr).put("property", tiles.get(locationu).get("property"));
-						prop.addIncludedDomino((Domino) tiles.get(locationr).get("domino"));
-
-					}
-				}
-				if (myUp && myDown) {
-					tile.put("property", tiles.get(locationu).get("property"));
-					Property prop = (Property) tiles.get(locationu).get("property");
-					prop.addIncludedDomino((Domino) tile.get("domino"));
-					if (!tiles.get(locationd).get("property").equals(tiles.get(locationu).get("property"))) {
-						Property tem = (Property) tiles.get(locationd).get("property");
-						tem.delete();
-						tiles.get(locationd).put("property", tiles.get(locationu).get("property"));
-						prop.addIncludedDomino((Domino) tiles.get(locationd).get("domino"));
-
-					}
-				}
-				if (myRight && myDown) {
+				
+				if (myLeft && myUp && myRight && myDown) {
+					
 					tile.put("property", tiles.get(locationr).get("property"));
 					Property prop = (Property) tiles.get(locationr).get("property");
 					prop.addIncludedDomino((Domino) tile.get("domino"));
+					if (!tiles.get(locationu).get("property").equals(tiles.get(locationr).get("property"))) {
+						Property tem = (Property) tiles.get(locationu).get("property");
+						tem.delete();
+						tiles.get(locationu).put("property", tiles.get(locationr).get("property"));
+						prop.addIncludedDomino((Domino) tiles.get(locationu).get("domino"));
+
+					}
+					if (!tiles.get(locationl).get("property").equals(tiles.get(locationr).get("property"))) {
+						Property tem = (Property) tiles.get(locationl).get("property");
+						tem.delete();
+						tiles.get(locationl).put("property", tiles.get(locationr).get("property"));
+						prop.addIncludedDomino((Domino) tiles.get(locationl).get("domino"));
+
+					}
 					if (!tiles.get(locationd).get("property").equals(tiles.get(locationr).get("property"))) {
 						Property tem = (Property) tiles.get(locationd).get("property");
 						tem.delete();
@@ -1546,6 +1497,7 @@ public class KingdominoController {
 
 					}
 				}
+				
 				if (myLeft && myUp && myRight) {
 					tile.put("property", tiles.get(locationr).get("property"));
 					Property prop = (Property) tiles.get(locationr).get("property");
@@ -1622,24 +1574,75 @@ public class KingdominoController {
 
 					}
 				}
-				if (myLeft && myUp && myRight && myDown) {
-					tile.put("property", tiles.get(locationr).get("property"));
-					Property prop = (Property) tiles.get(locationr).get("property");
+				if (myLeft && myRight) {
+					
+					tile.put("property", tiles.get(locationl).get("property"));
+					Property prop = (Property) tiles.get(locationl).get("property");
 					prop.addIncludedDomino((Domino) tile.get("domino"));
-					if (!tiles.get(locationu).get("property").equals(tiles.get(locationr).get("property"))) {
+					if (!tiles.get(locationr).get("property").equals(tiles.get(locationl).get("property"))) {
+						Property tem = (Property) tiles.get(locationr).get("property");
+						tem.delete();
+						tiles.get(locationr).put("property", tiles.get(locationl).get("property"));
+						prop.addIncludedDomino((Domino) tiles.get(locationr).get("domino"));
+
+					}
+				}
+				if (myLeft && myUp) {
+					
+					tile.put("property", tiles.get(locationl).get("property"));
+					Property prop = (Property) tiles.get(locationl).get("property");
+					prop.addIncludedDomino((Domino) tile.get("domino"));
+					if (!tiles.get(locationu).get("property").equals(tiles.get(locationl).get("property"))) {
 						Property tem = (Property) tiles.get(locationu).get("property");
 						tem.delete();
-						tiles.get(locationu).put("property", tiles.get(locationr).get("property"));
+						tiles.get(locationu).put("property", tiles.get(locationl).get("property"));
 						prop.addIncludedDomino((Domino) tiles.get(locationu).get("domino"));
 
 					}
-					if (!tiles.get(locationl).get("property").equals(tiles.get(locationr).get("property"))) {
-						Property tem = (Property) tiles.get(locationl).get("property");
+				}
+				if (myLeft && myDown) {
+					
+					tile.put("property", tiles.get(locationl).get("property"));
+					Property prop = (Property) tiles.get(locationl).get("property");
+					prop.addIncludedDomino((Domino) tile.get("domino"));
+					if (!tiles.get(locationd).get("property").equals(tiles.get(locationl).get("property"))) {
+						Property tem = (Property) tiles.get(locationd).get("property");
 						tem.delete();
-						tiles.get(locationl).put("property", tiles.get(locationr).get("property"));
-						prop.addIncludedDomino((Domino) tiles.get(locationl).get("domino"));
+						tiles.get(locationd).put("property", tiles.get(locationl).get("property"));
+						prop.addIncludedDomino((Domino) tiles.get(locationd).get("domino"));
 
 					}
+				}
+				if (myUp && myRight) {
+					tile.put("property", tiles.get(locationu).get("property"));
+					Property prop = (Property) tiles.get(locationu).get("property");
+					prop.addIncludedDomino((Domino) tile.get("domino"));
+					if (!tiles.get(locationr).get("property").equals(tiles.get(locationu).get("property"))) {
+						Property tem = (Property) tiles.get(locationr).get("property");
+						tem.delete();
+						tiles.get(locationr).put("property", tiles.get(locationu).get("property"));
+						prop.addIncludedDomino((Domino) tiles.get(locationr).get("domino"));
+
+					}
+				}
+				if (myUp && myDown) {
+					
+					tile.put("property", tiles.get(locationu).get("property"));
+					Property prop = (Property) tiles.get(locationu).get("property");
+					prop.addIncludedDomino((Domino) tile.get("domino"));
+					if (!tiles.get(locationd).get("property").equals(tiles.get(locationu).get("property"))) {
+						Property tem = (Property) tiles.get(locationd).get("property");
+						tem.delete();
+						tiles.get(locationd).put("property", tiles.get(locationu).get("property"));
+						prop.addIncludedDomino((Domino) tiles.get(locationd).get("domino"));
+
+					}
+				}
+				if (myRight && myDown) {
+					
+					tile.put("property", tiles.get(locationr).get("property"));
+					Property prop = (Property) tiles.get(locationr).get("property");
+					prop.addIncludedDomino((Domino) tile.get("domino"));
 					if (!tiles.get(locationd).get("property").equals(tiles.get(locationr).get("property"))) {
 						Property tem = (Property) tiles.get(locationd).get("property");
 						tem.delete();
@@ -1648,6 +1651,8 @@ public class KingdominoController {
 
 					}
 				}
+				
+				
 				if (myLeft) {
 
 					tile.put("property", tiles.get(locationl).get("property"));
@@ -2162,44 +2167,83 @@ public class KingdominoController {
 	// in order to know the winner of a finished game
 
 	public static void calculateRanking(Kingdomino kingdomino) {
-		ArrayList<Player> players = new ArrayList<Player>();
-		for (Player p : kingdomino.getCurrentGame().getPlayers()) {
-			players.add(p);
-		}
+		ArrayList<Player> players = new ArrayList<Player>(kingdomino.getCurrentGame().getPlayers());
+////		System.out.print(kingdomino.getCurrentGame().getPlayers());
+//		//sort players in arraylist based on totalscore
+//		Collections.sort(players, (a, b) -> b.getTotalScore() - a.getTotalScore());
+//		
+//		//set ranking based on position of above arraylist. does not consider resolvetiebreak yet.
+//		for(int i =0;i<players.size();i++) {
+//			for(int j = 0;j<players.size();j++) {
+//				if(kingdomino.getCurrentGame().getPlayer(i).equals(players.get(j))) {
+//					kingdomino.getCurrentGame().getPlayer(i).setCurrentRanking(j+1);
+//				}
+//			}
+//		}
+//		for(int i = 0;i<players.size();i++) {
+//			System.out.println("########");
+//			System.out.println(players.get(i).getKingdom().getProperties());
+//			System.out.println("########");
+//		}
+//		System.out.print(players);
+		
 		Player currentPlayer;
 		Player tempPlayer;
+		boolean tied = true;
 		for (int i = 0; i < players.size() - 1; i++) {
 			currentPlayer = players.get(i);
 			for (int j = i + 1; j < players.size(); j++) {
 				tempPlayer = players.get(j);
 				if (tempPlayer.getTotalScore() == currentPlayer.getTotalScore()) {
-					System.out.println("jdjdjdjdjdjd");
+					
 					switch (resolveTiebreak(tempPlayer, currentPlayer)) {
 					case 1:
+						
 						players.set(i, tempPlayer);
 						players.set(j, currentPlayer);
 						currentPlayer = tempPlayer;
+						tied=false;
 						break;
 					case 2:
 						players.set(i, tempPlayer);
 						players.set(j, currentPlayer);
 						currentPlayer = tempPlayer;
+						tied=false;
 						break;
 					case -1:
 						players.set(i, tempPlayer);
 						players.set(j, currentPlayer);
 						currentPlayer = tempPlayer;
+						tied=false;
 						break;
-					}
-				} else if (tempPlayer.getTotalScore() > currentPlayer.getTotalScore()) {
+				} }else if (tempPlayer.getTotalScore() > currentPlayer.getTotalScore()) {
 					players.set(i, tempPlayer);
 					players.set(j, currentPlayer);
 					currentPlayer = tempPlayer;
 				}
 			}
 		}
+		int tie = -1;
+		for (int i = 1; i < players.size(); i++) {
+			if (players.get(i).getTotalScore() == players.get(i - 1).getTotalScore() && tied) {
+				tie = i - 1;
+			}
+		}
+		boolean checked = false;
 		for (int i = 0; i < players.size(); i++) {
-			players.get(i).setCurrentRanking(i + 1);
+
+			if (tie == i) {
+				players.get(i).setCurrentRanking(i + 1);
+				players.get(i + 1).setCurrentRanking(i + 1);
+				checked = true;
+				i++;
+			}
+			if (checked) {
+				players.get(i).setCurrentRanking(i);
+			} else {
+				players.get(i).setCurrentRanking(i + 1);
+			}
+
 		}
 
 	}
